+++ src/haloray-core/simulation/crystalPopulationRepository.h
@@ -11,7 +11,7 @@ class CrystalPopulationRepository
 public:
     CrystalPopulationRepository();
     void add(CrystalPopulationPreset preset = Random);
-    void add(CrystalPopulation population, unsigned int weight, std::string name);
+    void add(CrystalPopulation population, double weight, std::string name);
     void remove(unsigned int index);
     void clear();
 
@@ -20,8 +20,8 @@ public:
     void setName(unsigned int index, std::string name);
     std::string getName(unsigned int index) const;
 
-    unsigned int getWeight(unsigned int index) const;
-    void setWeight(unsigned int index, unsigned int weight);
+    double getWeight(unsigned int index) const;
+    void setWeight(unsigned int index, double weight);
     double getProbability(unsigned int index) const;
 
     unsigned int getCount() const;
@@ -29,7 +29,7 @@ public:
 private:
     void addDefaults();
     std::vector<CrystalPopulation> m_crystals;
-    std::vector<unsigned int> m_weights;
+    std::vector<double> m_weights;
     std::vector<std::string> m_names;
     unsigned int m_nextId;
 };
